Refactoring of #309
