+++ src/main/java/spark/webserver/SparkServer.java
@@ -41,6 +41,8 @@ import org.eclipse.jetty.websocket.servlet.WebSocketCreator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import spark.webserver.websocket.WebSocketCreatorFactory;
+
 /**
  * Spark server implementation
  *
@@ -65,19 +67,19 @@ public class SparkServer {
      * Ignites the spark server, listening on the specified port, running SSL secured with the specified keystore
      * and truststore.  If truststore is null, keystore is reused.
      *
-     * @param host                    The address to listen on
-     * @param port                    - the port
-     * @param keystoreFile            - The keystore file location as string
-     * @param keystorePassword        - the password for the keystore
-     * @param truststoreFile          - the truststore file location as string, leave null to reuse keystore
-     * @param truststorePassword      - the trust store password
-     * @param staticFilesFolder       - the route to static files in classPath
-     * @param externalFilesFolder     - the route to static files external to classPath.
-     * @param latch                   - the countdown latch
-     * @param maxThreads              - max nbr of threads.
-     * @param minThreads              - min nbr of threads.
-     * @param threadIdleTimeoutMillis - idle timeout (ms).
-     * @param webSockerIdleTimeoutMIllis - Optional WebSocket idle timeout (ms).
+     * @param host                       The address to listen on
+     * @param port                       - the port
+     * @param keystoreFile               - The keystore file location as string
+     * @param keystorePassword           - the password for the keystore
+     * @param truststoreFile             - the truststore file location as string, leave null to reuse keystore
+     * @param truststorePassword         - the trust store password
+     * @param staticFilesFolder          - the route to static files in classPath
+     * @param externalFilesFolder        - the route to static files external to classPath.
+     * @param latch                      - the countdown latch
+     * @param maxThreads                 - max nbr of threads.
+     * @param minThreads                 - min nbr of threads.
+     * @param threadIdleTimeoutMillis    - idle timeout (ms).
+     * @param webSocketIdleTimeoutMillis - Optional WebSocket idle timeout (ms).
      */
     public void ignite(String host,
                        int port,
@@ -92,7 +94,7 @@ public class SparkServer {
                        int minThreads,
                        int threadIdleTimeoutMillis,
                        Map<String, Class<?>> webSocketHandlers,
-                       Optional<Integer> webSockerIdleTimeoutMillis) {
+                       Optional<Integer> webSocketIdleTimeoutMillis) {
 
         if (port == 0) {
             try (ServerSocket s = new ServerSocket(0)) {
@@ -123,23 +125,23 @@ public class SparkServer {
         server = connector.getServer();
         server.setConnectors(new Connector[] {connector});
 
-	ServletContextHandler webSocketServletContextHandler = null;
-	if (webSocketHandlers != null) {
-	    try {
-		webSocketServletContextHandler = new ServletContextHandler(null, "/", true, false);
-		WebSocketUpgradeFilter wsfilter = WebSocketUpgradeFilter.configureContext(webSocketServletContextHandler);
-		if (webSockerIdleTimeoutMillis.isPresent()) {
-		    wsfilter.getFactory().getPolicy().setIdleTimeout(webSockerIdleTimeoutMillis.get());
-		}
-		for (String path : webSocketHandlers.keySet()) {
-		    WebSocketCreator wscreator = WebSocketCreatorFactory.create(webSocketHandlers.get(path));
-		    wsfilter.addMapping(new ServletPathSpec(path), wscreator);
-		}
-	    } catch (Exception ex) {
-		logger.error("ignite failed", ex);
-		System.exit(100); // NOSONAR
-	    }
-	}
+        ServletContextHandler webSocketServletContextHandler = null;
+        if (webSocketHandlers != null) {
+            try {
+                webSocketServletContextHandler = new ServletContextHandler(null, "/", true, false);
+                WebSocketUpgradeFilter wsfilter = WebSocketUpgradeFilter.configureContext(webSocketServletContextHandler);
+                if (webSocketIdleTimeoutMillis.isPresent()) {
+                    wsfilter.getFactory().getPolicy().setIdleTimeout(webSocketIdleTimeoutMillis.get());
+                }
+                for (String path : webSocketHandlers.keySet()) {
+                    WebSocketCreator wscreator = WebSocketCreatorFactory.create(webSocketHandlers.get(path));
+                    wsfilter.addMapping(new ServletPathSpec(path), wscreator);
+                }
+            } catch (Exception ex) {
+                logger.error("ignite failed", ex);
+                System.exit(100); // NOSONAR
+            }
+        }
 
         // Handle static file routes
         if (staticFilesFolder == null && externalFilesFolder == null && webSocketServletContextHandler == null) {
